Split functional tests
